changed to float

+++ GithubPresentation/src/githubbbbb/SimpleCalculator.java
@@ -35,8 +35,8 @@ public class SimpleCalculator implements CalculatorInterface{
 	}
 	
 	@Override
-	public int modulo(int num1, int num2) {
-		int result = num1 % num2;
+	public float modulo(float num1, float num2) {
+		float result = num1 % num2;
 		return result;
 	}	
 }
